move: 컴포넌트 폴더 생성, 파일이동
