+++ src/components/lessonPage/Agreement.tsx
